fix: cleaned up urls for going to a blog post. Moved blog folder
