+++ pages/posts/blog/[id].tsx
@@ -1,7 +1,7 @@
-import Layout from "../../components/Layout";
-import { getAllPostIds, getPostData } from "../../lib/posts";
+import Layout from "../../../components/Layout";
+import { getAllPostIds, getPostData } from "../../../lib/posts";
 import Head from "next/head";
-import Date from "../../components/date";
+import Date from "../../../components/date";
 
 // const Post = () => {
 //   const router = useRouter();
